refactor: merge axelar-executable into axelar-solana-gateway
