+++ README.md
@@ -41,7 +41,6 @@ cargo xtask --help
 
 
 #### Utility crates
-- [**Axelar Executable**](solana/crates/axelar-executable/README.md): A set of libraries & interfaces that the destination program (3rd party integration) must implement.
 - [**Axelar Solana Encoding**](solana/crates/axelar-solana-encoding/README.md): Encoding used by the Multisig Prover to encode the data in a way that the relayer & the Solana Gateway can interpret.
 - [**Gateway Event Stack**](solana/crates/gateway-event-stack): The Relayer uses this crate to parse Gas Service & Gateway events.
