--- crates/axelar-executable/Cargo.toml
@@ -1,50 +1,0 @@
-[package]
-name = "axelar-executable"
-description = "Defines helpers for easily integrating with the Solana Axelar Gateway"
-version = "0.1.0"
-edition = "2021"
-
-[lints]
-workspace = true
-
-[lib]
-crate-type = ["cdylib", "lib"]
-
-
-[dependencies]
-axelar-solana-gateway = { workspace = true, default-features = false }
-borsh.workspace = true
-solana-program.workspace = true
-axelar-solana-encoding.workspace = true
-num-traits.workspace = true
-num-derive.workspace = true
-alloy-sol-types.workspace = true
-thiserror.workspace = true
-
-[dev-dependencies]
-evm-contracts-rs.workspace = true
-evm-contracts-test-suite.workspace = true
-rstest.workspace = true
-test-log.workspace = true
-tokio.workspace = true
-axelar-solana-gateway-test-fixtures.workspace = true
-rand.workspace = true
-
-[features]
-devnet-amplifier = [
-    "axelar-solana-gateway-test-fixtures/devnet-amplifier",
-    "axelar-solana-gateway/devnet-amplifier"
-]
-stagenet = [
-    "axelar-solana-gateway-test-fixtures/stagenet",
-    "axelar-solana-gateway/stagenet"
-]
-testnet = [
-    "axelar-solana-gateway-test-fixtures/testnet",
-    "axelar-solana-gateway/testnet"
-]
-mainnet = [
-    "axelar-solana-gateway-test-fixtures/mainnet",
-    "axelar-solana-gateway/mainnet"
-]
-default = ["devnet-amplifier"]
