+++ programs/axelar-solana-its/Cargo.toml
@@ -14,7 +14,6 @@ workspace = true
 no-entrypoint = []
 test-sbf = []
 devnet-amplifier = [
-    "axelar-executable/devnet-amplifier",
     "axelar-solana-gas-service/devnet-amplifier",
     "axelar-solana-gateway-test-fixtures/devnet-amplifier",
     "axelar-solana-gateway/devnet-amplifier",
@@ -22,7 +21,6 @@ devnet-amplifier = [
     "program-utils/devnet-amplifier"
 ]
 stagenet = [
-    "axelar-executable/stagenet",
     "axelar-solana-gas-service/stagenet",
     "axelar-solana-gateway-test-fixtures/stagenet",
     "axelar-solana-gateway/stagenet",
@@ -30,7 +28,6 @@ stagenet = [
     "program-utils/stagenet"
 ]
 testnet = [
-    "axelar-executable/testnet",
     "axelar-solana-gas-service/testnet",
     "axelar-solana-gateway-test-fixtures/testnet",
     "axelar-solana-gateway/testnet",
@@ -38,7 +35,6 @@ testnet = [
     "program-utils/testnet"
 ]
 mainnet = [
-    "axelar-executable/mainnet",
     "axelar-solana-gas-service/mainnet",
     "axelar-solana-gateway-test-fixtures/mainnet",
     "axelar-solana-gateway/mainnet",
@@ -50,7 +46,6 @@ default = ["devnet-amplifier"]
 [dependencies]
 alloy-primitives.workspace = true
 alloy-sol-types.workspace = true
-axelar-executable = { workspace = true, default-features = false }
 axelar-message-primitives.workspace = true
 axelar-solana-encoding.workspace = true
 axelar-solana-gas-service = { workspace = true, features = ["no-entrypoint"] }
