+++ programs/axelar-solana-memo-program/Cargo.toml
@@ -11,19 +11,16 @@ crate-type = ["cdylib", "lib"]
 [features]
 no-entrypoint = []
 devnet-amplifier = [
-    "axelar-executable/devnet-amplifier",
     "axelar-solana-gateway/devnet-amplifier",
     "axelar-solana-its/devnet-amplifier",
     "program-utils/devnet-amplifier"
 ]
 stagenet = [
-    "axelar-executable/stagenet",
     "axelar-solana-gateway/stagenet",
     "axelar-solana-its/stagenet",
     "program-utils/stagenet"
 ]
 testnet = [
-    "axelar-executable/testnet",
     "axelar-solana-gateway/testnet",
     "axelar-solana-its/testnet",
     "program-utils/testnet"
@@ -31,13 +28,11 @@ testnet = [
 mainnet = [
     "axelar-solana-gateway/mainnet",
     "axelar-solana-its/mainnet",
-    "axelar-executable/mainnet",
     "program-utils/mainnet"
 ]
 default = ["devnet-amplifier"]
 
 [dependencies]
-axelar-executable = { workspace = true, default-features = false }
 axelar-solana-encoding.workspace = true
 axelar-solana-gateway = { workspace = true, features = ["no-entrypoint"] }
 axelar-solana-its = { workspace = true, features = ["no-entrypoint"] }
