+++ programs/axelar-solana-multicall/Cargo.toml
@@ -14,25 +14,21 @@ workspace = true
 no-entrypoint = []
 test-sbf = []
 devnet-amplifier = [
-    "axelar-executable/devnet-amplifier",
     "axelar-solana-gateway-test-fixtures/devnet-amplifier",
     "axelar-solana-gateway/devnet-amplifier",
     "axelar-solana-memo-program/devnet-amplifier"
 ]
 stagenet = [
-    "axelar-executable/stagenet",
     "axelar-solana-gateway-test-fixtures/stagenet",
     "axelar-solana-gateway/stagenet",
     "axelar-solana-memo-program/stagenet",
 ]
 testnet = [
-    "axelar-executable/testnet",
     "axelar-solana-gateway-test-fixtures/testnet",
     "axelar-solana-gateway/testnet",
     "axelar-solana-memo-program/testnet",
 ]
 mainnet = [
-    "axelar-executable/mainnet",
     "axelar-solana-gateway-test-fixtures/mainnet",
     "axelar-solana-gateway/mainnet",
     "axelar-solana-memo-program/mainnet",
@@ -41,7 +37,6 @@ default = ["devnet-amplifier"]
 
 [dependencies]
 alloy-sol-types.workspace = true
-axelar-executable = { workspace = true, default-features = false }
 borsh.workspace = true
 axelar-solana-gateway = { workspace = true, features = ["no-entrypoint"] }
 solana-program.workspace = true
