+++ programs/axelar-solana-multicall/tests/module/from_evm_to_solana.rs
@@ -1,4 +1,4 @@
-use axelar_executable::AxelarMessagePayload;
+use axelar_solana_gateway::executable::AxelarMessagePayload;
 use axelar_solana_gateway_test_fixtures::gateway::random_message;
 use axelar_solana_memo_program::instruction::AxelarMemoInstruction;
 use axelar_solana_memo_program::state::Counter;
